ddapi: add testutil and new tests

wip: squash into last
