+++ go/pkg/ddapi/handler_test.go
@@ -15,10 +15,7 @@
 package ddapi
 
 import (
-	"bytes"
 	"context"
-	"fmt"
-	"io"
 	"io/ioutil"
 	"net/http"
 	"net/http/httptest"
@@ -26,31 +23,150 @@ import (
 	"regexp"
 	"strings"
 	"testing"
-	"text/template"
-	"time"
 
 	log "github.com/sirupsen/logrus"
 	"github.com/stretchr/testify/assert"
 	"github.com/stretchr/testify/suite"
 	"github.com/testcontainers/testcontainers-go"
-	"github.com/testcontainers/testcontainers-go/wait"
 )
 
 /*
-Shared these resources between tests:
+Share these resources between tests:
 - a containerized Cortex instance
 - a DDCortexProxy instance
 
 Use `testify/suite` primitives for managing setup and teardown of these
 resources.
 
-Ref:
-https://godoc.org/github.com/stretchr/testify/suite
+Start with the test definitions: Test* methods on Suite struct, and put the
+boilerplate code (struct definition, setup/teardown) towards the bottom of the
+file
+
+Each suite method whose name starts with `Test` is run as a regular test.
+
+Ref: https://godoc.org/github.com/stretchr/testify/suite
 
 Interesting recipe:
 https://github.com/stretchr/testify/pull/655#issuecomment-588500729
 
 */
+
+func (suite *Suite) TestCortexDirectly() {
+	// Test interacting straight with the containerized Cortex Expect 405
+	// response: Method Not Allowed (only POST is supposed to work). In that
+	// sense, this test explicitly checks availability of Cortex to other
+	// tests.
+	resp, err := http.Get(suite.cortexPushURL)
+	assert.NoError(suite.T(), err)
+	assert.Equal(suite.T(), resp.StatusCode, 405)
+}
+
+func (suite *Suite) TestPostMissingCTH() {
+
+	// Valid JSON in body, valid URL, valid method. Invalid: missing
+	// content-type header.
+	req := httptest.NewRequest(
+		"POST",
+		"http://localhost/api/v1/series",
+		strings.NewReader("{}"),
+	)
+
+	w := httptest.NewRecorder()
+	suite.ddcp.SeriesPostHandler(w, req)
+
+	checker := func(w *httptest.ResponseRecorder) {
+		resp := w.Result()
+		assert.Equal(suite.T(), 400, resp.StatusCode)
+
+		rbody, err := ioutil.ReadAll(resp.Body)
+		if err != nil {
+			suite.T().Errorf("readAll error: %v", err)
+		}
+
+		// Confirm that the original error message (for why the request could
+		// not be proxied) is contained in the response body.
+		assert.Equal(
+			suite.T(),
+			"bad request: request lacks content-type header",
+			strings.TrimSpace(string(rbody)),
+		)
+	}
+
+	checker(w)
+}
+
+func (suite *Suite) TestPostEmptyBody() {
+	// Valid URL, valid method, valid CT header, invalid: missing body
+	req := httptest.NewRequest("POST", "http://localhost/api/v1/series", nil)
+	req.Header.Set("Content-Type", "application/json")
+	w := httptest.NewRecorder()
+	suite.ddcp.SeriesPostHandler(w, req)
+
+	checker := func(w *httptest.ResponseRecorder) {
+		resp := w.Result()
+
+		// Expect bad request, because there's no JSON body in the request.
+		assert.Equal(suite.T(), 400, resp.StatusCode)
+
+		rbody, err := ioutil.ReadAll(resp.Body)
+		if err != nil {
+			suite.T().Errorf("readAll error: %v", err)
+		}
+
+		// Confirm that the original error message (for why the request could
+		// not be proxied) is contained in the response body.
+		assert.Regexp(
+			suite.T(),
+			regexp.MustCompile("^bad request: error while translating body: invalid JSON doc: readObjectStart: expect .*$"),
+			strings.TrimSpace(string(rbody)),
+		)
+	}
+
+	checker(w)
+}
+
+func (suite *Suite) TestPostSimpleBody() {
+	testname := suite.T().Name()
+	w := httptest.NewRecorder()
+
+	suite.ddcp.SeriesPostHandler(
+		w,
+		genSubmitRequest(createJSONBody(testname)),
+	)
+
+	checker := func(w *httptest.ResponseRecorder) {
+		resp := w.Result()
+		assert.Equal(suite.T(), 202, resp.StatusCode)
+
+		rbody, err := ioutil.ReadAll(resp.Body)
+		if err != nil {
+			suite.T().Errorf("readAll error: %v", err)
+		}
+
+		assert.Equal(
+			suite.T(),
+			"{\"status\": \"ok\"}",
+			strings.TrimSpace(string(rbody)),
+		)
+	}
+
+	checker(w)
+}
+
+// Generate HTTP request to POST to the DD API proxy, specifically to
+// <baseURL>/api/v1/series with valid header(s) and valid URL, so that
+// individual tests do not need to repeat code for constructing a 'good' HTTP
+// request.
+func genSubmitRequest(jsontext string) *http.Request {
+	req := httptest.NewRequest(
+		"POST",
+		"http://localhost/api/v1/series",
+		strings.NewReader(jsontext),
+	)
+	req.Header.Set("Content-Type", "application/json")
+	log.Infof("generated POST request with body:\n%s", jsontext)
+	return req
+}
 
 type Suite struct {
 	suite.Suite
@@ -107,210 +223,7 @@ func (suite *Suite) TearDownSuite() {
 	}
 }
 
-// Each suite method whose name starts with `Test` is run as a regular test.
-func (suite *Suite) TestCortexDirectly() {
-	// Test interacting straight with the containerized Cortex Expect 405
-	// response: Method Not Allowed (only POST is supposed to work). In that
-	// sense, this test explicitly checks availability of Cortex to other
-	// tests.
-	resp, err := http.Get(suite.cortexPushURL)
-	assert.NoError(suite.T(), err)
-	assert.Equal(suite.T(), resp.StatusCode, 405)
-}
-
-func (suite *Suite) TestPostEmptyBody() {
-	req := httptest.NewRequest("POST", "http://localhost/api/v1/series", nil)
-	w := httptest.NewRecorder()
-	suite.ddcp.SeriesPostHandler(w, req)
-
-	checker := func(w *httptest.ResponseRecorder) {
-		resp := w.Result()
-
-		// Expect bad request, because there's no JSON body in the request
-		// And no content type header.
-		assert.Equal(suite.T(), 400, resp.StatusCode)
-
-		rbody, err := ioutil.ReadAll(resp.Body)
-		if err != nil {
-			suite.T().Errorf("readAll error: %v", err)
-		}
-
-		// Confirm that the original error message (for why the request could
-		// not be proxied) is contained in the response body.
-		assert.Regexp(
-			suite.T(),
-			regexp.MustCompile("^bad request: error while translating body: invalid JSON doc: readObjectStart: expect .*$"),
-			strings.TrimSpace(string(rbody)))
-	}
-
-	checker(w)
-}
-
-func createJSONBody(metricname string) string {
-	const jsonTextTemplate = `
-	{
-		"series": [
-		  {
-			"metric": "unit.test.metric.{{.metricname}}",
-			"points": [
-			  [
-				{{.secondsSinceEpoch}},
-				1
-			  ]
-			],
-			"tags": [
-			  "version:7.24.1"
-			],
-			"host": "x1carb6",
-			"type": "rate",
-			"interval": 10
-		  }
-		]
-	  }
-`
-	data := map[string]interface{}{
-		"metricname":        metricname,
-		"secondsSinceEpoch": fmt.Sprintf("%v", time.Now().Unix()),
-	}
-
-	// Render template, store result in string.
-
-	t := template.Must(template.New("").Parse(jsonTextTemplate))
-
-	buf := &bytes.Buffer{}
-	if err := t.Execute(buf, data); err != nil {
-		panic(err)
-	}
-
-	// Return buffer (type *bytes.Buffer), can be passed as third argument to
-	// httptest.NewRequest(). This buffer holds UTF-8-encoded JSON. return buf
-
-	// Return string (JSON doc as text). Need to create strings.NewReader(s)
-	// then for passing it as third arg to httptest.NewRequest()
-	return buf.String()
-}
-
-func (suite *Suite) TestPostSimpleBody() {
-	jsonText := createJSONBody("dummymetric")
-	log.Infof("POST:\n%s", jsonText)
-	req := httptest.NewRequest(
-		"POST",
-		"http://localhost/api/v1/series",
-		strings.NewReader(jsonText),
-	)
-
-	w := httptest.NewRecorder()
-	suite.ddcp.SeriesPostHandler(w, req)
-
-	checker := func(w *httptest.ResponseRecorder) {
-		resp := w.Result()
-		assert.Equal(suite.T(), 202, resp.StatusCode)
-
-		rbody, err := ioutil.ReadAll(resp.Body)
-		if err != nil {
-			suite.T().Errorf("readAll error: %v", err)
-		}
-		log.Infof("response body:\n%s", rbody)
-		// Confirm that the original error message (for why the request could
-		// not be proxied) is contained in the response body.
-		assert.Equal(
-			suite.T(),
-			"{\"status\": \"ok\"}",
-			strings.TrimSpace(string(rbody)),
-		)
-	}
-
-	checker(w)
-}
-
-// In order for 'go test' to run this suite, create a normal test function.
-func TestSuite(t *testing.T) {
+// For 'go test' to run this suite, create a stdlib `testing` test function.
+func TestWithCortexSuite(t *testing.T) {
 	suite.Run(t, new(Suite))
-}
-
-func createTmpDir() (string, error) {
-	// Create a unique temporary directory for this test suite run. Note: "It
-	// is the caller's responsibility to remove the directory when no longer
-	// needed." -- do this in suite teardown.
-	tempdir, err := ioutil.TempDir("/tmp", "opstrace-go-unit-tests")
-	if err != nil {
-		return "", err
-	}
-
-	log.Infof("created suite tmp dir: %s", tempdir)
-	return tempdir, nil
-}
-
-func startCortex(tempdir string) (context.Context, testcontainers.Container, string, error) {
-	ctx := context.Background()
-
-	// Assume that this is the directory that this source file resides in.
-	// That's true when clicking `run test` in vs code, and interestingly also
-	// true when running `make unit-tests`, i.e. when running `go test ...`.
-	// https://stackoverflow.com/a/23847429/145400
-	testdir, err := os.Getwd()
-	log.Infof("test dir: %s", testdir)
-	if err != nil {
-		log.Fatal(err)
-	}
-
-	// Copy the Cortex config file into `tempdir` (under /tmp) so that it can
-	// so that the bindMount source is valid _on the host_, too (assume that
-	// this test runner runs in a container, and that /tmp is shared between
-	// host and containers).
-	cortexCfgSourcePath := testdir + "/test-resources/cortex-dev-cfg.yaml"
-	cortexCfgHostPath := tempdir + "/cortex-dev-cfg.yaml"
-	simpleFileCopy(cortexCfgSourcePath, cortexCfgHostPath)
-
-	req := testcontainers.ContainerRequest{
-		Image:        "cortexproject/cortex:v1.6.0",
-		ExposedPorts: []string{"33333/tcp"},
-		WaitingFor:   wait.ForHTTP("/ready").WithPort("33333/tcp"),
-		BindMounts:   map[string]string{cortexCfgHostPath: "/cortex-config.yaml"},
-		Cmd:          []string{"exec", "./cortex", "-config.file=/cortex-config.yaml"},
-	}
-
-	container, err := testcontainers.GenericContainer(ctx, testcontainers.GenericContainerRequest{
-		ContainerRequest: req,
-		Started:          true,
-	})
-
-	if err != nil {
-		return nil, nil, "", err
-	}
-
-	// Construct (base) URL to remote_write endpoint
-	ip, err := container.Host(ctx)
-	if err != nil {
-		panic(err)
-	}
-	port, err := container.MappedPort(ctx, "33333")
-	if err != nil {
-		panic(err)
-	}
-	pushURL := fmt.Sprintf("http://%s:%s/api/v1/push", ip, port.Port())
-
-	return ctx, container, pushURL, nil
-}
-
-// Copy the src file to dst. Any existing file will be overwritten and will not
-// copy file attributes. https://stackoverflow.com/a/21061062/145400
-func simpleFileCopy(src, dst string) error {
-	in, err := os.Open(src)
-	if err != nil {
-		return err
-	}
-	defer in.Close()
-
-	out, err := os.Create(dst)
-	if err != nil {
-		return err
-	}
-	defer out.Close()
-
-	_, err = io.Copy(out, in)
-	if err != nil {
-		return err
-	}
-	return out.Close()
 }
